+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-29 20:20:54
-// Commit: fix(perf): fix memory leak by cancelling perf subscription
+// Updated at: 2026-05-30 13:44:05
+// Commit: feat(perf): add bi-directional event bus handler for perf
 
 class perfComponentState {
-  final String commitRef = '1601';
-  final String lastMessage = 'fix(perf): fix memory leak by cancelling perf subscription';
-  final DateTime timestamp = DateTime.parse('2026-05-29T20:20:54+05:00');
+  final String commitRef = '1619';
+  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-30T13:44:05+05:00');
 }